+++ src/main/java/org/hv/biscuits/spine/identify/BisGenerationType.java
@@ -1,0 +1,11 @@
+package org.hv.biscuits.spine.identify;
+
+/**
+ * @author wujianchuan
+ */
+public interface BisGenerationType {
+    /**
+     * 服务标识+表标识+站标识+年月日+序列号
+     */
+    String SERVER_TABLE_STATION_DATE_INCREMENT = "serverTableStationDateIncrement";
+}