+++ src/main/java/org/hv/biscuits/spine/model/PostUserRelation.java
@@ -1,14 +1,14 @@
 package org.hv.biscuits.spine.model;
 
+import org.hv.biscuits.spine.AbstractBisEntity;
 import org.hv.pocket.annotation.Column;
 import org.hv.pocket.annotation.Entity;
-import org.hv.pocket.model.BaseEntity;
 
 /**
  * @author leyan95 2020/6/26
  */
-@Entity(table = "T_POST_USER", tableId = 117)
-public class PostUserRelation extends BaseEntity {
+@Entity(table = "T_POST_USER")
+public class PostUserRelation extends AbstractBisEntity {
     private static final long serialVersionUID = -1878385856874316067L;
     @Column
     private String postUuid;