+++ src/main/java/org/hv/biscuits/spine/model/UserRoleRelation.java
@@ -1,14 +1,14 @@
 package org.hv.biscuits.spine.model;
 
+import org.hv.biscuits.spine.AbstractBisEntity;
 import org.hv.pocket.annotation.Entity;
 import org.hv.pocket.annotation.ManyToOne;
-import org.hv.pocket.model.BaseEntity;
 
 /**
  * @author leyan95 2019/1/30
  */
-@Entity(table = "T_USER_ROLE", tableId = 107)
-public class UserRoleRelation extends BaseEntity {
+@Entity(table = "T_USER_ROLE")
+public class UserRoleRelation extends AbstractBisEntity {
     private static final long serialVersionUID = 509835395399250891L;
     @ManyToOne(columnName = "USER_UUID", clazz = User.class, upBridgeField = "uuid")
     private String userUuid;